- Summary: fix bug126686: Support anonymous/unknown report output format - Bugzilla Bug (s) Resolved: 126686 - Description: When set output format other "html" and "pdf", use correct logic to determine a proper renderer. - Tests Description  :
- Files Edited:

- Files Added:

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

+++ chart/org.eclipse.birt.chart.reportitem/src/org/eclipse/birt/chart/reportitem/ChartReportStyleProcessor.java
@@ -136,7 +136,7 @@ public class ChartReportStyleProcessor implements IStyleProcessor
 			ta.setHorizontalAlignment( ha );
 			ta.setVerticalAlignment( va );
 			FontDefinition fd = FontDefinitionImpl.create( fname,
-					(float) fsize,
+					fsize,
 					fbold,
 					fitalic,
 					funder,
